% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable.R
\name{airtable}
\alias{airtable}
\title{Create a new airtable object}
\usage{
airtable(
  table,
  base,
  view = NULL,
  api_url = "https://api.airtable.com",
  api_version = 0
)
}
\arguments{
\item{table}{Table name}

\item{base}{Airtable base containing table}

\item{view}{Optional view of data to read}

\item{api_url}{API endpoint to connect to. Can be changed for integrations that require custom endpoint}

\item{api_version}{Version of API to use}
}
\value{
An airtable object
}
\description{
Create a new airtable object
}
\examples{
\dontrun{
table <- airtable("Table 1", "appXXXXXXXXXXXXX")
}
}
