% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable_base.R
\name{airtable_base}
\alias{airtable_base}
\title{Get metadata for an Airtable base}
\usage{
airtable_base(base, metadata_api_url_pattern = deprecated(), token = NULL)
}
\arguments{
\item{base}{A valid Airtable base ID}

\item{metadata_api_url_pattern}{URL for GET request. Defaults to
"meta/bases/{base_id}/tables"}

\item{token}{Airtable personal access token}
}
\value{
an \code{airtable_base_schema} object
}
\description{
Get metadata for an Airtable base.
}
\examples{
\dontrun{
base <- airtable_base("appXXXXXXXXXXXXX")

# read 'table1' from the base
tbl_data <- read_airtable(base$table1)

# view base schema
print(base$schema)
}
}
