% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable_fields_schema.R
\name{airtable_fields_schema}
\alias{airtable_fields_schema}
\title{Create an `airtable_fields_schema` object that can be used for data validation}
\usage{
airtable_fields_schema(field_schema)
}
\arguments{
\item{field_schema}{A list of lists. Each child list should have a `name`, `type`, and `id` as well as an optional `options`}
}
\value{
An `airtable_fields_schema`object
}
\description{
Create an `airtable_fields_schema` object that can be used for data validation
}
