% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_airtable.R, R/utils_httr2.R
\name{req_airtable_records}
\alias{req_airtable_records}
\alias{airtable_request}
\alias{req_airtable_query}
\alias{req_auth_airtable}
\title{Use httr2 to create an Airtable API request}
\usage{
req_airtable_records(
  req = NULL,
  url = NULL,
  api_url = NULL,
  api_version = NULL,
  base = NULL,
  table = NULL,
  record = NULL,
  view = NULL,
  sort = NULL,
  max_records = 100,
  page_size = NULL,
  tz = NULL,
  locale = NULL,
  fields_by_id = FALSE,
  fields = NULL,
  filter = NULL,
  desc = FALSE,
  cell_format = NULL,
  token = NULL,
  default = "AIRTABLE_PAT"
)

airtable_request(
  url = NULL,
  api_url = NULL,
  api_version = NULL,
  base = NULL,
  table = NULL,
  airtable = NULL,
  ...,
  call = caller_env()
)

req_airtable_query(
  .req = NULL,
  ...,
  template = NULL,
  method = NULL,
  data = NULL,
  url = NULL,
  api_url = NULL,
  api_version = NULL,
  airtable = NULL,
  token = NULL,
  string = NULL
)

req_auth_airtable(
  req,
  token = NULL,
  default = "AIRTABLE_PAT",
  string = NULL,
  rate = 5/1,
  realm = NULL
)
}
\arguments{
\item{req, .req}{A \link[httr2]{request}.}

\item{url}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \link{airtable_request() } returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any additional value provided to url is ignored. For
the metadata API where only a base ID is required include \code{require_base = TRUE} to pass to \code{check_airtable_api_url()}.}

\item{api_url}{Airtable API URL, If \code{NULL} (default), the api_url is set to
getOption("rairtable.api_url", "https://api.airtable.com").}

\item{api_version}{Airtable API version number, If \code{NULL} (default), the
api_version is set to getOption("rairtable.api_version", 0)).}

\item{base}{Airtable base id starting with with "app".}

\item{table}{Airtable table id or name. If table is a table ID it is a string
starting with "viw".}

\item{record}{Airtable record identifier, Default: \code{NULL}}

\item{view}{Airtable view identifier, Default: \code{NULL}}

\item{sort}{Field to sort by, Default: \code{NULL}}

\item{max_records}{Maximum number of records to return, Default: \code{NULL}. Must
be 100 or less.}

\item{tz, locale}{Time zone and locale, Defaults: \code{NULL}}

\item{fields_by_id}{If \code{TRUE}, return fields by id, Default: \code{FALSE}}

\item{fields}{Fields to return from Airtable base, Default: \code{NULL}}

\item{filter}{Filter to apply to records, Note: This parameter is a
placeholder and is not currently implemented. Default: \code{NULL}}

\item{desc}{If \code{TRUE}, sort in descending order, Default: \code{FALSE}}

\item{cell_format}{Cell format for "Link to another record" fields (either
"json" (unique ID) or "string" (displayed character string)), Default:
'json'}

\item{token}{Airtable personal access token, Default: \code{NULL} (set to output
from \code{get_airtable_pat_or_key()})}

\item{default}{Default name for environmental variable for token.}

\item{...}{
  Arguments passed on to \code{\link[=check_airtable_api_url]{check_airtable_api_url}}
  \describe{
    \item{\code{require_base}}{If \code{TRUE}, the string "tbl" must be included in the
supplied url.}
    \item{\code{require_table}}{If \code{TRUE}, the string "tbl" must be included in the
supplied url. If a string with a table name is provided, the string must be
part of the string supplied to url. If \code{FALSE} (default), there is no check
for a table being present in the URL.}
    \item{\code{require_view}}{If \code{TRUE}, the string "viw" must be included in the
string supplied to url.}
  }}

\item{template}{Template for query parameters passed to
\code{\link[httr2:req_template]{httr2::req_template()}}, Default: \code{NULL}.}

\item{method}{Custom HTTP method}

\item{data}{Data to include in body.}

\item{string}{Passed to \code{\link[httr2:req_user_agent]{httr2::req_user_agent()}}, Default: \code{NULL} (set to
\code{getOption("rairtable.useragent", default = "rairtable (https://github.com/matthewjrogers/rairtable)")})}

\item{rate}{Rate passed to \code{\link[httr2:req_throttle]{httr2::req_throttle()}}. Defaults to \code{5 / 1}}

\item{realm}{Passed to \code{\link[httr2:req_throttle]{httr2::req_throttle()}}. Defaults to \code{NULL}.}

\item{per_page}{Max records to return per page, Default: \code{NULL}}
}
\description{
\code{\link[=airtable_request]{airtable_request()}} creates an initial request object based on an airtable
object, API url, Airtable view URL, or base and table IDs.
}
\details{
\code{\link[=req_auth_airtable]{req_auth_airtable()}} set the rate limit, user agent, and authenticates the
Airtable API request with an API key or personal access token.

\code{\link[=req_airtable_query]{req_airtable_query()}} builds an API request optionally using a template or
custom HTTP method (other than GET or POST). This function can create an
initial request using a URL or airtable object or use an existing request
object. The function also calls \code{\link[=req_auth_airtable]{req_auth_airtable()}} before returning the
request. Most of the other data access functions in this package wrap
\code{\link[=req_auth_airtable]{req_auth_airtable()}}.
}
\keyword{internal}
