% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable_url.R
\name{airtable_url}
\alias{airtable_url}
\alias{is_airtable_url}
\alias{is_airtable_api_url}
\alias{check_airtable_url}
\alias{check_airtable_api_url}
\alias{parse_airtable_url}
\title{Test, check, and parse Airtable URLs and API URLs}
\usage{
is_airtable_url(url, base_url = NULL)

is_airtable_api_url(url, api_url = NULL)

check_airtable_url(url, base_url = NULL, call = caller_env())

check_airtable_api_url(
  url,
  require_base = TRUE,
  require_table = FALSE,
  require_view = FALSE,
  api_url = NULL,
  call = caller_env()
)

parse_airtable_url(
  url,
  base_url = NULL,
  api_url = NULL,
  table_name = NULL,
  view_name = NULL,
  call = caller_env()
)
}
\arguments{
\item{url}{Airtable API or view URL, typically in the form
"https://api.airtable.com/v0/{baseId}/{tableIdOrName}" or
"https://airtable.com/{baseID}/{tableIdOrName}/{viewId}"}

\item{base_url}{Base URL for an Airtable base or view. Defaults \code{NULL} and
set to \code{getOption("rairtable.api_url", "https://airtable.com/")}.}

\item{api_url}{Expected base URL for an API url. Defaults to \code{NULL} and set
to \code{getOption("rairtable.api_url", "https://api.airtable.com")}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{require_base}{If \code{TRUE}, the string "tbl" must be included in the
supplied url.}

\item{require_table}{If \code{TRUE}, the string "tbl" must be included in the
supplied url. If a string with a table name is provided, the string must be
part of the string supplied to url. If \code{FALSE} (default), there is no check
for a table being present in the URL.}

\item{require_view}{If \code{TRUE}, the string "viw" must be included in the
string supplied to url.}

\item{table_name, view_name}{Pattern to use when parsing a table name or view
name. Defaults to \code{NULL} which extracts strings starting with "tbl" for
tables IDs and "viw" for view IDs.}
}
\description{
Utility functions for checking and validating Airtable URLs and API URLs:
}
\details{
\itemize{
\item \code{\link[=is_airtable_url]{is_airtable_url()}} returns \code{TRUE} if a url starts with the base_url value.
\item \code{\link[=is_airtable_api_url]{is_airtable_api_url()}} returns \code{TRUE} if a url starts with the api_url value.
\item \code{\link[=check_airtable_url]{check_airtable_url()}} errors if url does not start with the base_url value.
\item \code{\link[=check_airtable_api_url]{check_airtable_api_url()}} errors if url is not a valid API url.
\item \code{\link[=parse_airtable_url]{parse_airtable_url()}} accepts a base or API url as an input and returns a
named list with the base, table, and view ID values.
}

For more information on the expected URL format for different Airtable API
end points, see the web API documentation at
\url{https://airtable.com/developers/web/api/introduction}
}
