% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_get}
\alias{api_get}
\title{Get a subset of records via the API}
\usage{
api_get(airtable, max_records)
}
\arguments{
\item{airtable}{An object of class `airtable`}

\item{max_records}{Maximum number of records to return}
}
\description{
Get a subset of records via the API
}
