% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batch_encode_patch}
\alias{batch_encode_patch}
\title{JSON encode a dataframe for PATCH}
\usage{
batch_encode_patch(df, id_col = NULL)
}
\arguments{
\item{df}{A data frame}

\item{id_col}{Optional column name containing IDs as a string}
}
\value{
JSON object
}
\description{
JSON encode a dataframe for PATCH
}
