% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batch_encode_post}
\alias{batch_encode_post}
\title{JSON encode a dataframe for POST}
\usage{
batch_encode_post(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
JSON object
}
\description{
JSON encode a dataframe for POST
}
