% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_url.R
\name{check_airtable_url}
\alias{check_airtable_url}
\alias{check_airtable_api_url}
\title{Check Airtable URLs and API URLs}
\usage{
check_airtable_url(
  url,
  base_url = NULL,
  allow_null = FALSE,
  allow_shared = FALSE,
  call = caller_env()
)

check_airtable_api_url(
  url,
  require_base = TRUE,
  require_table = FALSE,
  require_view = FALSE,
  api_url = NULL,
  call = caller_env()
)
}
\arguments{
\item{url}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \code{\link[=request_airtable]{request_airtable()}} returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any supplied url is ignored.}

\item{allow_null}{If \code{FALSE} (default) and url is \code{NULL},
\code{\link[=check_airtable_url]{check_airtable_url()}} errors. If \code{TRUE}, a \code{NULL} value for url is
allowed.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{require_base}{If \code{TRUE} (default), the string "app" must be included in
the supplied url.}

\item{require_table}{If \code{TRUE}, the string "tbl" must be included in the
supplied url. If require_table is a string, it is treated as the required
table name which must be part of url. If \code{FALSE} (default), there is no
check for a table being present in the URL.}

\item{require_view}{If \code{TRUE}, the string "viw" must be included in the
string supplied to url. If \code{FALSE} (default), there is no check
for a view being present in the URL.}

\item{api_url}{Airtable API URL, If \code{NULL} (default), the api_url is set to
\code{getOption("rairtable.api_url", "https://api.airtable.com")}.}
}
\description{
Utility functions for checking and validating Airtable URLs and API URLs.
}
\details{
\itemize{
\item \code{\link[=check_airtable_url]{check_airtable_url()}} errors if url does not start with the base_url value.
\item \code{\link[=check_airtable_api_url]{check_airtable_api_url()}} errors if url is not a valid API url.
}

For more information on the expected URL format for different Airtable API
end points, see the web API documentation at
\url{https://airtable.com/developers/web/api/introduction}
}
\keyword{internal}
