% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_names}
\alias{compare_names}
\title{Compare the names of two data frames and stop execution if they do not match}
\usage{
compare_names(df1, df2)
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{A dataframe}
}
\description{
Compare the names of two data frames and stop execution if they do not match
}
