% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_base.R
\name{create_base}
\alias{create_base}
\title{Create a new Airtable base}
\usage{
create_base(
  name,
  workspace = NULL,
  tables = NULL,
  schema = NULL,
  simplifyVector = FALSE,
  token = NULL,
  ...
)
}
\arguments{
\item{workspace}{A workspace ID (a string starting with "wsp") or URL
containing a workspace ID.}

\item{tables}{An array of table config lists. Data frames are not currently
supported.}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{...}{Additional parameters passed to \code{\link[=req_airtable]{req_airtable()}}}
}
\description{
Use the Airtable Metadata API to create a new Airtable bases. If schema is
provided and tables is \code{NULL}, tables is copied based on the models from an
existing base schema.
}
\seealso{
\itemize{
\item \code{\link[=get_base_schema]{get_base_schema()}}
\item \code{\link[=copy_table_config]{copy_table_config()}}
}
}
