% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\alias{update_table}
\title{Create or update an Airtable table}
\usage{
create_table(
  airtable = NULL,
  config = NULL,
  name = NULL,
  description = NULL,
  fields = NULL,
  model = NULL,
  token = NULL,
  ...
)

update_table(airtable = NULL, name = NULL, description = NULL, ...)
}
\arguments{
\item{airtable}{An airtable class object created with \code{\link[=airtable]{airtable()}}. airtable
must include a request_url attribute. url is ignored if airtable is
provided. Optional if url or base \emph{and} table are supplied.}

\item{config}{A table model with a name, description, and field array from
\code{\link[=make_table_config]{make_table_config()}}.}

\item{name}{Table name. Optional for \code{\link[=copy_table_config]{copy_table_config()}} unless the table
config is expected to be used within the same base.}

\item{description}{Table description. Optional.}

\item{fields}{Field list.}

\item{model}{A model for an existing table to copy for a new table config. If
name or description are supplied, any existing name or description from the
model is ignored. model is ignored if table is supplied.}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{...}{Additional parameters:
\itemize{
\item For \code{\link[=airtable_base]{airtable_base()}} and \code{\link[=get_base_schema]{get_base_schema()}}, additional parameters
may include airtable or url (one of which required if base is \code{NULL}).
\item For \code{\link[=get_table_models]{get_table_models()}}, additional parameters are passed to
\code{\link[=get_base_schema]{get_base_schema()}} so may include simplifyVector.
\item For \code{\link[=get_table_model]{get_table_model()}}, additional parameters can include airtable or url
(one of which is required if table is \code{NULL}).
}}
}
\value{
Invisibly returns a list with the table ID, name, description,
primary field ID, and field names, types, and IDs.
}
\description{
Create or update an Airtable table or create a table model. \code{\link[=create_table]{create_table()}}
requires both a name and fields (or a model with a name and fields). Either a
name or description must be provided for \code{\link[=update_table]{update_table()}}.
}
\details{
Find more information on the create table API:
\url{https://airtable.com/developers/web/api/create-table} or the update table
API: \url{https://airtable.com/developers/web/api/update-table} Find more
information about the table config for the Airtable API:
\url{https://airtable.com/developers/web/api/model/table-model}
}
\examples{
\dontrun{
if (interactive()) {
  create_table(
    base = "<exmaple base ID>",
    name = "Apartments",
    description = "A to-do list of places to visit",
    fields = list(
      list(
        "description" = "Name of the apartment",
        "name" = "Name",
        "type" = "singleLineText"
      ),
      list(
        "name" = "Address",
        "type" = "singleLineText"
      ),
      list(
        "name" = "Visited",
        "options" = list(
          "color" = "greenBright",
          "icon" = "check"
        ),
        "type" = "checkbox"
      )
    )
  )
}
}
}
