% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_funs.R
\name{delete}
\alias{delete}
\title{Send batch delete request}
\usage{
delete(ids, airtable_obj)
}
\arguments{
\item{ids}{Vector of IDs}

\item{airtable_obj}{Object of class `airtable`}
}
\description{
Send batch delete request
}
