% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{download_attachment_field}
\alias{download_attachment_field}
\title{Download files from an Airtable attachment field}
\usage{
download_attachment_field(
  data,
  field = NULL,
  path = NULL,
  ...,
  call = caller_env()
)
}
\arguments{
\item{data}{An Airtable data frame with an attachment list column selected
with the field parameter \emph{or} an attachment list column.}

\item{field}{Attachment field or column name as string or tidyselect function.}

\item{path}{Path to download file.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Learn more about the attachment field type:
\url{https://airtable.com/developers/web/api/field-model#multipleattachment}

Learn more about attachments in Airtable:
\url{https://support.airtable.com/docs/airtable-attachment-url-behavior}
}
\keyword{internal}
