% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_types.R
\docType{data}
\name{field_types}
\alias{field_types}
\title{Airtable field types}
\format{
An object of class \code{character} of length 33.
}
\usage{
field_types
}
\description{
A vector of Airtable field types as of 2023-04-18. Find more information on
Airtable field types at
\url{https://airtable.com/developers/web/api/model/field-type}
}
\keyword{datasets}
