% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_airtable_pat.R
\name{get_airtable_pat_or_key}
\alias{get_airtable_pat_or_key}
\title{Get an Airtable Personal Access Token or an API Key (if PAT is not available)}
\usage{
get_airtable_pat_or_key(
  token = NULL,
  default = c("AIRTABLE_PAT", "AIRTABLE_API_KEY"),
  call = caller_env()
)
}
\description{
Get an Airtable Personal Access Token or an API Key (if PAT is not available)
}
\keyword{internal}
