% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ids}
\alias{get_ids}
\title{Title}
\usage{
get_ids(df, id_col)
}
\arguments{
\item{df}{A dataframe}

\item{id_col}{Optional name of a column containing Airtable record IDs}
}
\value{
Vector of Airtable IDs
}
\description{
Title
}
