% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_records}
\alias{get_records}
\alias{get_record_id_col}
\title{Get records IDs from a data.frame column, data.frame row names, or a
character vector}
\usage{
get_records(
  data = NULL,
  id_col = NULL,
  id_col_arg = caller_arg(id_col),
  records = NULL,
  call = caller_env()
)

get_record_id_col(
  data,
  id_col = NULL,
  id_col_arg = caller_arg(id_col),
  call = caller_env()
)
}
\description{
Get records IDs from a data.frame column, data.frame row names, or a
character vector
}
\keyword{internal}
