% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_config.R
\name{get_table_model}
\alias{get_table_model}
\title{Get a table schema from a base schema or table name}
\usage{
get_table_model(
  table = NULL,
  schema = NULL,
  simplifyVector = FALSE,
  ...,
  call = caller_env()
)
}
\arguments{
\item{table}{A base schema or, if schema is provided, a string with a table
name or ID.}

\item{schema}{A base schema list from \code{\link[=get_base_schema]{get_base_schema()}} using
\code{simplifyVector = FALSE}. Optional if base is supplied.}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{...}{Additional parameters passed to \code{\link[=get_table_models]{get_table_models()}}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
An internal function used by \code{\link[=copy_table_config]{copy_table_config()}}.
}
\keyword{internal}
