% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable_url.R
\name{is_airtable_api_url}
\alias{is_airtable_api_url}
\title{Does url match the pattern of an Airtable API url?}
\usage{
is_airtable_api_url(url, api_url = NULL)
}
\arguments{
\item{url}{Airtable URL or API URL, typically in the form
"https://airtable.com/{baseID}/{tableIdOrName}/{viewId}" or
"https://api.airtable.com/v0/{baseId}/{tableIdOrName}". Note that, by
default, shared URLs of the form "https://airtable.com/shr..." are not
supported.}

\item{api_url}{Expected base URL for an API url. Defaults to \code{NULL} and set
to \code{getOption("rairtable.api_url", "https://api.airtable.com/v0")}.}
}
\description{
\code{\link[=is_airtable_api_url]{is_airtable_api_url()}} returns \code{TRUE} if a url starts with the api_url
value.
}
\keyword{internal}
