% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable.R
\name{is_airtable_obj}
\alias{is_airtable_obj}
\alias{check_airtable_obj}
\title{Test or check if an object is an airtable object}
\usage{
is_airtable_obj(x)

check_airtable_obj(
  x,
  require_url = TRUE,
  require_table = TRUE,
  require_view = FALSE,
  require_fields = FALSE,
  call = caller_env()
)
}
\arguments{
\item{x}{Object to test or check.}

\item{require_url}{If \code{TRUE} (default), check that x includes a "request_url"
attribute that is a valid Airtable API url.}

\item{require_table}{If \code{TRUE}, the string "tbl" must be included in the
supplied url. If a string with a table name is provided, the string must be
part of the string supplied to url. If \code{FALSE} (default), there is no check
for a table being present in the URL.}

\item{require_view}{If \code{TRUE}, the string "viw" must be included in the
string supplied to url.}

\item{require_fields}{If \code{TRUE}, check that x includes a "fields" value that
is a list with the \code{airtable_fields_schema} class. Defaults to \code{FALSE}.}
}
\description{
Test or check if an object is an airtable object
}
\keyword{internal}
