% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bases.R
\name{list_bases}
\alias{list_bases}
\title{List Airtable bases accessible with a personal access token}
\usage{
list_bases(bases = NULL, token = NULL, simplifyVector = TRUE, ...)
}
\arguments{
\item{bases}{A character vectors of base ID or name values. If provided, the
results of list_bases are filtered to matching bases only. Defaults to
\code{NULL} which returns all bases associated with the personal access token or
API key.}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{...}{Additional parameters passed to \code{\link[=req_airtable]{req_airtable()}}}
}
\value{
A tibble or list of base IDs, names, and permission levels or a list
if \code{simplifyVector = FALSE}.
}
\description{
Use the Airtable Metadata API to list all Airtable bases accessible with a
personal access token. Optionally filter bases by name or base ID.
}
\seealso{
\code{\link[=get_table_models]{get_table_models()}}
}
