% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_comments.R
\name{list_comments}
\alias{list_comments}
\alias{create_comment}
\alias{delete_comment}
\title{List comments, create a comment, or delete a comment on an Airtable record}
\usage{
list_comments(airtable = NULL, record = NULL, ...)

create_comment(airtable = NULL, record = NULL, text, ...)

delete_comment(airtable = NULL, record = NULL, comment, ...)
}
\arguments{
\item{airtable}{An airtable class object created with \code{\link[=airtable]{airtable()}}. airtable
must include a request_url attribute. url is ignored if airtable is
provided. Optional if url or base \emph{and} table are supplied.}

\item{record}{Record ID (a string starting with "rec") or an Airtable record
URL. If record is a URL, no additional parameters are required. Otherwise,
airtable, url, or base and table must be supplied.}

\item{...}{Additional parameters including url or base and table values. url
or base and table must be supplied if airtable is \code{NULL} and record is a
record ID.}

\item{text}{String with text of comment to create. Required for
\code{\link[=create_comment]{create_comment()}}.}

\item{comment}{Comment ID (string starting with the text "com"). Required for
\code{\link[=delete_comment]{delete_comment()}}.}
}
\value{
\itemize{
\item \code{\link[=list_comments]{list_comments()}} returns a data frame with record comment IDs, text,
and author information.
\item \code{\link[=create_comment]{create_comment()}} returns a data frame with the created comment ID, text,
and author information.
\item \code{\link[=delete_comment]{delete_comment()}} returns a list with a comment ID..
}
}
\description{
List all comments for an Airtable record, create a new comment, or delete an
existing comment on an Airtable record.
}
