% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_config.R
\name{make_table_config}
\alias{make_table_config}
\alias{copy_table_config}
\title{Make a new table config or copy a table config from an existing table model}
\usage{
make_table_config(
  name = NULL,
  description = NULL,
  fields = NULL,
  model = NULL,
  ...,
  call = caller_env()
)

copy_table_config(
  table = NULL,
  model = NULL,
  name = NULL,
  description = NULL,
  ...,
  call = caller_env()
)
}
\arguments{
\item{name}{Table name. Optional for \code{\link[=copy_table_config]{copy_table_config()}} unless the table
config is expected to be used within the same base.}

\item{description}{Table description. Optional.}

\item{fields}{Field list.}

\item{model}{A model for an existing table to copy for a new table config. If
name or description are supplied, any existing name or description from the
model is ignored. model is ignored if table is supplied.}

\item{...}{
  Arguments passed on to \code{\link[=get_table_model]{get_table_model}}
  \describe{
    \item{\code{schema}}{A base schema list from \code{\link[=get_base_schema]{get_base_schema()}} using
\code{simplifyVector = FALSE}. Optional if base is supplied.}
    \item{\code{simplifyVector}}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}
  }}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{table}{A base schema or, if schema is provided, a string with a table
name or ID.}
}
\description{
Note that copying a config from a table model requires modifying or ignoring
some field types. See \code{\link[=copy_field_config]{copy_field_config()}} for details.

Learn more about the Airtable Web API table config object:
\url{https://airtable.com/developers/web/api/model/table-config}
}
