% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_airtable}
\alias{new_airtable}
\title{Airtable constructor}
\usage{
new_airtable(
  table = character(),
  base = character(),
  view = character(),
  api_url = character(),
  api_version = integer()
)
}
\arguments{
\item{table}{Table name}

\item{base}{base name}

\item{view}{view name}

\item{api_url}{api url}

\item{api_version}{api version}
}
\value{
Airtable object
}
\description{
Airtable constructor
}
