% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_config.R
\name{prep_table_field}
\alias{prep_table_field}
\title{Prep a field configuration from a table model for use as a new field config}
\usage{
prep_table_field(field, linkedTableId = NULL, viewIdForRecordSelection = NULL)
}
\arguments{
\item{linkedTableId}{Table ID for a linked table option to use with any
multipleRecordLinks field type.}

\item{viewIdForRecordSelection}{View ID from linked table to use for record
selection on multipleRecordLinks field type.}
}
\description{
Internal helper function.
}
\keyword{internal}
