% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_error}
\alias{process_error}
\title{Return an informative error response}
\usage{
process_error(response_status)
}
\arguments{
\item{response_status}{httr response status}
}
\value{
An informative error message
}
\description{
Return an informative error response
}
