% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{rairtable-config}
\alias{rairtable-config}
\title{rairtable environment variables and options}
\description{
rairtable environment variables and options
}
\section{User facing configuration}{


These are environment variables and options that users may set, to modify the behavior of rairtable.
\subsection{User facing environment variables}{
\subsection{\code{rairtable.id_col}}{

A number of functions use the \code{airtable_id_col} parameter to set the name of the column holding record id values. The default value "airtable_id_col" can be modified using the \code{rairtable.id_col} option. This option is used by \code{\link[=read_airtable]{read_airtable()}}, \code{\link[=update_records]{update_records()}}, \code{\link[=delete_records]{delete_records()}}, and other functions.
}

\subsection{\code{rairtable.user_agent}}{

The rairtable package uses \code{\link[httr2:req_user_agent]{httr2::req_user_agent()}} to pass a User-Agent header with each API call. The default value "rairtable (https://github.com/matthewjrogers/rairtable)" can be modified using the  \code{rairtable.user_agent} option.
}

\subsection{\code{rairtable.safely}}{

The default value for \code{rairtable.safely} (\code{TRUE}) requires user confirmation every time the user deletes, modifies, or creates records for Airtable. This can be modified using the \code{rairtable.safely} option.
}

\subsection{\code{rairtable.fields-trunc}}{

This option is passed to the style parameter of \code{\link[cli:cli_vec]{cli::cli_vec()}} to determine the number of fields printed for \code{airtable} and \code{airtable_table_schema} objects. Option defaults to 6.
}

}

\subsection{API-specific environment variables}{

The API environment variables should remain the same unless the Airtable API changes.
\subsection{\code{rairtable.batch_size}}{

Airtable limits the number of records that can be deleted, updated, or created in a single API call to 10. This default value can be modified using the \code{rairtable.batch_size} option.
}

\subsection{\code{rairtable.api_url}}{

The base URL for the Airtable API is "https://api.airtable.com". This default value can be modified using the \code{rairtable.api_url} option. This option is used by internal functions for checking Airtable API urls and, optionally, \code{\link[=request_airtable]{request_airtable()}}.
}

\subsection{\code{rairtable.api_version}}{

The Airtable API version number is \code{0}. This default value can be modified using
the \code{rairtable.api_version} option. This option may be used by
\code{\link[=request_airtable]{request_airtable()}}.
}

\subsection{\code{rairtable.base_url}}{

The Airtable view base URL is "http://airtable.com". This default value can be modified using the \code{rairtable.base_url} option. This option is used by \code{\link[=is_airtable_url]{is_airtable_url()}} and \code{\link[=check_airtable_url]{check_airtable_url()}}.
}

}
}

