% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rairtable-package.R
\docType{package}
\name{rairtable-package}
\alias{rairtable}
\alias{rairtable-package}
\title{rairtable: Efficient Wrapper for the Airtable Web API}
\description{
Create, read, update, and delete Airtable records or read base metadata using the Airtable Web API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matthewjrogers.github.io/rairtable/}
  \item Report bugs at \url{https://github.com/matthewjrogers/rairtable/issues}
}

}
\author{
\strong{Maintainer}: Matthew Rogers \email{matthew.rogers09@gmail.com}

Authors:
\itemize{
  \item Eli Pousson (\href{https://orcid.org/0000-0001-8280-1706}{ORCID})
}

}
\keyword{internal}
