% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable_schema.R
\name{req_airtable_schema}
\alias{req_airtable_schema}
\title{Create a request for the Airtable metadata API}
\usage{
req_airtable_schema(
  url = NULL,
  base = NULL,
  table = NULL,
  column = NULL,
  airtable = NULL,
  ...,
  type = c("schema", "list", "update", "create"),
  from = c("base", "table", "field"),
  token = NULL,
  call = caller_env()
)
}
\arguments{
\item{url}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \link{airtable_request() } returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any additional value provided to url is ignored. For
the metadata API where only a base ID is required include \code{require_base = TRUE} to pass to \code{check_airtable_api_url()}.}

\item{base}{Airtable base id starting with with "app".}

\item{table}{Airtable table id or name. If table is a table ID it is a string
starting with "viw".}

\item{column}{Column ID, Defaults to \code{NULL}. Only required when type =
"update" and from = "field".}

\item{...}{PARAM_DESCRIPTION}

\item{type}{Type of API call to use, Default: c("schema", "list", "update", "create")}

\item{from}{Is the API call for bases, tables, or fields? Default: c("base",
"table", "field"). Not all combinations of type and from are supported.
"base" works with "schema", "create", and "list", "table" and "field" both
only work with "update" and "create".}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A modified HTTP request.
}
\description{
Create a request to get metadata or update metadata for bases, tables, or
fields using the Airtable metadata API.
}
\keyword{internal}
