% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{req_airtable_url}
\alias{req_airtable_url}
\alias{req_airtable_query}
\alias{req_auth_airtable}
\alias{req_airtable_records}
\title{Use httr2 to create and perform an Airtable API request}
\usage{
req_airtable_url(
  url = NULL,
  api_url = NULL,
  api_version = NULL,
  base = NULL,
  table = NULL
)

req_airtable_query(
  .req = NULL,
  ...,
  template = NULL,
  url = NULL,
  api_url = NULL,
  api_version = NULL,
  token = NULL,
  string = NULL
)

req_auth_airtable(
  req,
  token = NULL,
  default = "AIRTABLE_PAT",
  string = NULL,
  rate = 5/1,
  realm = NULL
)

req_airtable_records(
  req = NULL,
  url = NULL,
  api_url = NULL,
  api_version = NULL,
  base = NULL,
  table = NULL,
  record = NULL,
  view = NULL,
  sort = NULL,
  max_records = 100,
  page_size = NULL,
  tz = NULL,
  locale = NULL,
  fields_by_id = FALSE,
  fields = NULL,
  filter = NULL,
  desc = FALSE,
  cell_format = NULL,
  token = NULL,
  default = "AIRTABLE_PAT"
)
}
\arguments{
\item{url}{If url is provided and is a URL, return \code{httr2::request(url)}.}

\item{api_url}{Airtable API URL, Default: \code{NULL} (set to
getOption("rairtable.api_url", "https://api.airtable.com"))}

\item{api_version}{Airtable API version number, Default: \code{NULL}  (set to
getOption("rairtable.api_version", 0))}

\item{base}{Airtable base identifier. Required. If base is an Airtable URL,
the base, table, and view identifiers are extracted from the URL and
assigned to any \code{NULL} values for table and view. Required if req is
\code{NULL}. Ignored if req is provided.}

\item{table}{Airtable table name or identifier. Required if req is \code{NULL}.
Ignored if req is provided.}

\item{...}{Additional parameters passed to \code{\link[httr2:req_template]{httr2::req_template()}} if
template is not \code{NULL} or \code{\link[httr2:req_url]{httr2::req_url_query()}} if template is \code{NULL}.}

\item{template}{Template for query parameters passed to
\code{\link[httr2:req_template]{httr2::req_template()}}, Default: \code{NULL}.}

\item{token}{Airtable personal access token, Default: \code{NULL} (set to output
from \code{get_airtable_pat()})}

\item{string}{Passed to \code{\link[httr2:req_user_agent]{httr2::req_user_agent()}}, Default: \code{NULL} (set to
\code{getOption("rairtable.useragent", default = "rairtable (https://github.com/matthewjrogers/rairtable)")})}

\item{default}{Default name for environmental variable for token.}

\item{rate}{Rate passed to \code{\link[httr2:req_throttle]{httr2::req_throttle()}}. Defaults to \code{5 / 1}}

\item{realm}{Passed to \code{\link[httr2:req_throttle]{httr2::req_throttle()}}. Defaults to \code{NULL}.}

\item{record}{Airtable record identifier, Default: \code{NULL}}

\item{view}{Airtable view identifier, Default: \code{NULL}}

\item{sort}{Field to sort by, Default: \code{NULL}}

\item{max_records}{Maximum number of records to return, Default: \code{NULL}. Must
be 100 or less.}

\item{tz, locale}{Time zone and locale, Defaults: \code{NULL}}

\item{fields_by_id}{If \code{TRUE}, return fields by id, Default: \code{FALSE}}

\item{fields}{Fields to return from Airtable base, Default: \code{NULL}}

\item{filter}{Filter to apply to records, Note: This parameter is a
placeholder and is not currently implemented. Default: \code{NULL}}

\item{desc}{If \code{TRUE}, sort in descending order, Default: \code{FALSE}}

\item{cell_format}{Cell format for "Link to another record" fields (either
"json" (unique ID) or "string" (displayed character string)), Default:
'json'}

\item{per_page}{Max records to return per page, Default: \code{NULL}}
}
\description{
\code{\link[=req_auth_airtable]{req_auth_airtable()}} supports the API request by setting rate limit, user
agent, and authenticate Airtable request with key/token
}
\keyword{internal}
