% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_airtable.R
\name{req_auth_airtable}
\alias{req_auth_airtable}
\title{Set the rate limit, user agent, and authenticate an Airtable API request}
\usage{
req_auth_airtable(
  req,
  token = NULL,
  default = "AIRTABLE_PAT",
  string = NULL,
  rate = 5/1,
  realm = NULL,
  allow_key = TRUE
)
}
\arguments{
\item{req}{A request object created by \code{\link[httr2:request]{httr2::request()}} or
\code{\link[=request_airtable]{request_airtable()}}. Required.}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{default}{Default name for environmental variable for token.}

\item{string}{Passed to \code{\link[httr2:req_user_agent]{httr2::req_user_agent()}}, Default: \code{NULL} (set to
\code{getOption("rairtable.useragent", default = "rairtable (https://github.com/matthewjrogers/rairtable)")})}

\item{rate}{Rate passed to \code{\link[httr2:req_throttle]{httr2::req_throttle()}}. Defaults to \code{5 / 1}.}

\item{realm}{Passed to \code{\link[httr2:req_throttle]{httr2::req_throttle()}}. Defaults to \code{NULL}.}

\item{allow_key}{If \code{TRUE}, allow use of an Airtable API key or an Airtable
personal access token (PAT) to authenticate the request. The metadata API
does not support the API key so allow_key is set to \code{FALSE} for any
functions that call that API.}
}
\description{
Set the rate limit, user agent, and authenticates the Airtable API request
with an API key or personal access token.
}
\keyword{internal}
