% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_records.R
\name{req_create_records}
\alias{req_create_records}
\title{Use httr2 to make a POST request to create records in an Airtable table}
\usage{
req_create_records(
  req = NULL,
  ...,
  data,
  typecast = TRUE,
  token = NULL,
  call = caller_env()
)
}
\arguments{
\item{req}{A request object created by \code{\link[httr2:request]{httr2::request()}} or
\code{\link[=request_airtable]{request_airtable()}}. If .req is provided, any supplied url, api_url,
api_version, or airtable parameters are ignored.}

\item{...}{
  Arguments passed on to \code{\link[=request_airtable]{request_airtable}}
  \describe{
    \item{\code{url}}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \code{\link[=request_airtable]{request_airtable()}} returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any additional value provided to url is ignored. For
the metadata API (where only a base ID is required) include \code{require_base = TRUE} to pass to \code{\link[=check_airtable_api_url]{check_airtable_api_url()}} when creating a request.}
    \item{\code{base}}{Airtable base id starting with with "app". Optional if url or
airtable are supplied or if require_base is \code{FALSE}. base and table are
both required if url and airtable are \code{NULL} and require_base is \code{TRUE}.}
    \item{\code{table}}{Airtable table id or name. Table ID values are strings starting
with "tbl". Optional if require_table is \code{FALSE}.}
    \item{\code{view}}{Airtable view ID. View ID values starts with "viw". Optional if
require_view is \code{FALSE}.}
    \item{\code{airtable}}{An airtable class object created with \code{\link[=airtable]{airtable()}}. airtable
must include a request_url attribute. url is ignored if airtable is
provided.  Optional if url or base \emph{and} table are supplied.}
    \item{\code{require_base}}{If \code{TRUE} (default), the string "app" must be included in
the supplied url.}
    \item{\code{require_table}}{If \code{TRUE}, the string "tbl" must be included in the
supplied url. If a string with a table name is provided, the string must be
part of the string supplied to url. If \code{FALSE} (default), there is no check
for a table being present in the URL.}
    \item{\code{require_view}}{If \code{TRUE}, the string "viw" must be included in the
string supplied to url. If \code{FALSE} (default), there is no check
for a view being present in the URL.}
  }}

\item{data}{A data frame or list of records to create in Airtable. Column
names or the names of list values must match the field names used in the
Airtable base. Required.}

\item{typecast}{If \code{TRUE} (default), values will be converted to match the
base if possible. typecase must be set to \code{TRUE} to add new values to a
multiselect field type.
base. Required.}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
Invisibly return response from API call.
}
\description{
Use httr2 to make a POST request to create records in an Airtable table
}
\keyword{internal}
