% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_records.R
\name{req_delete_records}
\alias{req_delete_records}
\title{Delete one or more records from an Airtable table}
\usage{
req_delete_records(url = NULL, ..., records, batch = TRUE, call = caller_env())
}
\arguments{
\item{url}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \link{airtable_request() } returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any additional value provided to url is ignored. For
the metadata API where only a base ID is required include \code{require_base = TRUE} to pass to \code{check_airtable_api_url()}.}

\item{...}{
  Arguments passed on to \code{\link[=airtable_request]{airtable_request}}
  \describe{
    \item{\code{api_url}}{Airtable API URL, If \code{NULL} (default), the api_url is set to
getOption("rairtable.api_url", "https://api.airtable.com").}
    \item{\code{api_version}}{Airtable API version number, If \code{NULL} (default), the
api_version is set to getOption("rairtable.api_version", 0)).}
    \item{\code{base}}{Airtable base id starting with with "app".}
    \item{\code{table}}{Airtable table id or name. If table is a table ID it is a string
starting with "viw".}
  }}

\item{records}{Required. Character vector with record IDs. If only one record
is provided, the parameters are passed to \code{\link[=req_delete_record]{req_delete_record()}}
https://airtable.com/developers/web/api/delete-multiple-records.}

\item{batch}{If \code{TRUE} (default) and more than 10 records are provided, the
records vector is split into batches of 10. If batch is \code{FALSE}, records
must be length 10 or less. This default batch size is set by
"rairtable.batch_size" option. The option default is 10 may be changed if
the API limit is modified in the future.}
}
\description{
Delete one record from an Airtable table with \code{\link[=req_delete_record]{req_delete_record()}} or delete
multiple records with \code{\link[=req_delete_records]{req_delete_records()}}.
}
\details{
More information on the Airtable API:
\itemize{
\item \url{https://airtable.com/developers/web/api/delete-record}
\item \url{https://airtable.com/developers/web/api/delete-multiple-records}
}
}
\keyword{internal}
