% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_airtable.R
\name{req_list_records}
\alias{req_list_records}
\title{Build a request for the Airtable list records API method}
\usage{
req_list_records(
  req = NULL,
  airtable = NULL,
  view = NULL,
  sort = NULL,
  max_records = 100,
  page_size = NULL,
  tz = NULL,
  locale = NULL,
  fields_by_id = FALSE,
  fields = NULL,
  desc = FALSE,
  cell_format = NULL,
  token = NULL,
  ...,
  call = caller_env()
)
}
\arguments{
\item{view}{Airtable view ID or name, Default: \code{NULL}. If the supplied url or
airtable object includes a view, the view provided to this parameter is
ignored.}

\item{sort}{Field to sort by, Default: \code{NULL}}

\item{max_records}{Maximum number of records to return, Default: \code{NULL}. Must
be 100 or less.}

\item{page_size}{Max records to return per page, Default: \code{NULL}}

\item{tz, locale}{Time zone and locale, Defaults to \code{NULL}. If cell_format is
"string", tz defaults to \code{Sys.timezone()} and locale defaults to
\code{Sys.getlocale("LC_TIME")}.}

\item{fields_by_id}{If \code{TRUE}, return fields by id, Default: \code{FALSE}}

\item{fields}{Fields to return from Airtable base, Default: \code{NULL}}

\item{desc}{If \code{TRUE}, sort in descending order, Default: \code{FALSE}}

\item{cell_format}{Cell format for "Link to another record" fields. Defaults
to "json" which returns a unique record ID. A "string" cell_format returns
the displayed character string.}
}
\description{
Build a request for the Airtable list records API method
}
\keyword{internal}
