% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_airtable.R
\name{req_perform_offset}
\alias{req_perform_offset}
\alias{resp_body_records}
\title{Get the body from API requests with a series of offset values}
\usage{
req_perform_offset(
  req,
  offset = NULL,
  type = "combine",
  record_cols = c("id", "createdTime"),
  record_nm = record_cols,
  max_rows = NULL,
  call = caller_env()
)

resp_body_records(
  resp,
  simplifyVector = TRUE,
  type = "combine",
  record_cols = c("id", "createdTime"),
  record_nm = NULL,
  call = caller_env(),
  ...
)
}
\arguments{
\item{req}{A modified HTTP request from \code{\link[=airtable_request]{airtable_request()}} or a httr2
function.}

\item{offset}{Offset value to passed to \code{\link[httr2:req_url]{httr2::req_url_query()}} as part of
the API call. See
\url{https://airtable.com/developers/web/api/list-records#response-offset} for
more information on how the offset value is used by the Airtable API.}

\item{type}{"combine" (default) combines the fields and records data.frames
in the response. Additional supported options are "records" and "fields".}

\item{record_cols}{Column names for columns to retain from records data.frame
when type is "records" or "combine".}

\item{record_nm}{Names to use for additional columns indicated by
record_cols. Defaults to \code{NULL}.}

\item{id_col}{An alternate name to use for the id column of the response
data.frame.}
}
\description{
Get the body from API requests with a series of offset values
}
\keyword{internal}
