% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_records.R
\name{req_update_records}
\alias{req_update_records}
\alias{req_update_record}
\title{Update one or more records}
\usage{
req_update_records(
  url = NULL,
  ...,
  records,
  data,
  typecast = FALSE,
  method = NULL,
  call = caller_env()
)

req_update_record(
  url = NULL,
  ...,
  record,
  data,
  typecast = FALSE,
  method = NULL,
  call = caller_env()
)
}
\arguments{
\item{url}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \code{\link[=airtable_request]{airtable_request()}} returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any additional value provided to url is ignored. For
the metadata API (where only a base ID is required) include \code{require_base = TRUE} to pass to \code{\link[=check_airtable_api_url]{check_airtable_api_url()}} when creating a request.}

\item{...}{For \code{\link[=airtable_request]{airtable_request()}}, additional parameters passed to
\code{\link[=check_airtable_api_url]{check_airtable_api_url()}}. For \code{\link[=req_query_airtable]{req_query_airtable()}}, additional
parameters passed to \code{\link[httr2:req_template]{httr2::req_template()}} if template is not \code{NULL} or
\code{\link[httr2:req_url]{httr2::req_url_query()}} if template is \code{NULL}.}

\item{records, record}{Record ID or IDs to update as a list or character
vector.}

\item{data}{Data to include in body.}

\item{method}{Custom HTTP method}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
https://airtable.com/developers/web/api/update-record
https://airtable.com/developers/web/api/update-multiple-records
}
\keyword{internal}
