% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable_schema.R, R/request_airtable_meta.R
\name{request_airtable_meta}
\alias{request_airtable_meta}
\title{Create a request for the Airtable metadata API}
\usage{
request_airtable_meta(
  url = NULL,
  base = NULL,
  table = NULL,
  column = NULL,
  airtable = NULL,
  ...,
  meta,
  token = NULL,
  call = caller_env()
)

request_airtable_meta(
  url = NULL,
  base = NULL,
  table = NULL,
  column = NULL,
  airtable = NULL,
  ...,
  meta,
  token = NULL,
  call = caller_env()
)
}
\arguments{
\item{url}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \code{\link[=request_airtable]{request_airtable()}} returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any additional value provided to url is ignored. For
the metadata API (where only a base ID is required) include \code{require_base = TRUE} to pass to \code{\link[=check_airtable_api_url]{check_airtable_api_url()}} when creating a request.}

\item{base}{Airtable base id starting with with "app". Optional if url or
airtable are supplied or if require_base is \code{FALSE}. base and table are
both required if url and airtable are \code{NULL} and require_base is \code{TRUE}.}

\item{table}{Airtable table id or name. Table ID values are strings starting
with "tbl". Optional if require_table is \code{FALSE}.}

\item{column}{Column ID, Defaults to \code{NULL}. Only required when type =
"update" and from = "field".}

\item{airtable}{An airtable class object created with \code{\link[=airtable]{airtable()}}. airtable
must include a request_url attribute. url is ignored if airtable is
provided.  Optional if url or base \emph{and} table are supplied.}

\item{...}{Additional parameters passed to \code{\link[=req_airtable]{req_airtable()}}}

\item{meta}{Type of API call to use, Default: c("schema", "list", "update",
"create")}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{what}{Is the API call for bases, tables, or fields? Default: c("base",
"table", "field"). Not all combinations of type and from are supported.
"base" works with "schema", "create", and "list", "table" and "field" both
only work with "update" and "create".}
}
\value{
A modified HTTP request.

A modified HTTP request.
}
\description{
Create a request to get metadata or update metadata for bases, tables, or
fields using the Airtable metadata API. Note that the metadata API does not
support the legacy Airtable API key so a personal access token is required.
You can save a personal access token with \code{\link[=set_airtable_pat]{set_airtable_pat()}}.

Create a request to get metadata or update metadata for bases, tables, or
fields using the Airtable metadata API. Note that the metadata API does not
support the legacy Airtable API key so a personal access token is required.
You can save a personal access token with \code{\link[=set_airtable_pat]{set_airtable_pat()}}.
}
\keyword{internal}
