% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_data_resp}
\alias{return_data_resp}
\title{Return data, response body, or combined body from list of response objects}
\usage{
return_data_resp(data = NULL, resp = NULL, return_data = TRUE)
}
\arguments{
\item{data}{Input data frame or list}

\item{resp}{A response object.}

\item{return_data}{If \code{FALSE}, return JSON response from the Airtable web API
as a list. If \code{TRUE} (default) and data is not \code{NULL}, return input data
frame or list.}
}
\description{
An internal helper function to optionally return data, response body as a
list, or combined list of httr2_res response objects.
}
\keyword{internal}
