% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safety_check}
\alias{safety_check}
\title{If the user chooses to execute safely, confirm action before proceeding}
\usage{
safety_check(safely, cancel_message, ...)
}
\arguments{
\item{safely}{A boolean}

\item{cancel_message}{A message to send if cancellation is chosen}

\item{...}{One or morse expressions to print in the menu}
}
\description{
If the user chooses to execute safely, confirm action before proceeding
}
