% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_airtable_pat.R
\name{set_airtable_pat}
\alias{set_airtable_pat}
\title{Set or install Airtable personal access token}
\usage{
set_airtable_pat(pat, install = FALSE)
}
\arguments{
\item{pat}{A valid Airtable personal access token}

\item{install}{Add your personal access token to your .Renviron for future sessions. Optionally overwrite an existing Airtable personal access token.}
}
\value{
No return value, called for side effects
}
\description{
Set Airtable personal access token as an environment variable, and optionally install the personal access token to your .Renviron file for future use.
}
\examples{
\dontrun{
set_airtable_pat("XXXXXXXXXX", install = TRUE)
}

}
