% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_list}
\alias{split_list}
\title{Split list or vector into equal size pieces}
\usage{
split_list(lst, chunk_size = 10)
}
\arguments{
\item{lst}{A list or vector}

\item{chunk_size}{Maximum length of the pieces}
}
\value{
A list of with each element having length chunk_size or less
}
\description{
Split list or vector into equal size pieces
}
