% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_rows}
\alias{split_rows}
\title{Split a dataframe by row}
\usage{
split_rows(df, chunk_size)
}
\arguments{
\item{df}{A data frame}

\item{chunk_size}{Size of splits}
}
\value{
List
}
\description{
Split a dataframe by row
}
