% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_quietly}
\alias{stop_quietly}
\title{Exit function without error}
\usage{
stop_quietly(...)
}
\arguments{
\item{...}{Any number of objects to be concetenated into a message}
}
\description{
Exit function without error
}
