% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_records.R
\name{update_records}
\alias{update_records}
\title{Update Airtable records}
\usage{
update_records(
  data,
  airtable = NULL,
  airtable_id_col = NULL,
  columns = tidyselect::everything(),
  records = NULL,
  safely = TRUE,
  return_json = FALSE,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{data}{A data.frame containing the fields to update and, if records is
\code{NULL}, a column or rownames with record IDs.}

\item{airtable}{An airtable object. Optional if url or base and table are
provided with additional parameters passed to \code{\link[=airtable_request]{airtable_request()}}.}

\item{airtable_id_col}{Column containing Airtable record IDs. Not required if
record IDs are stored in row names as returned from \code{\link[=read_airtable]{read_airtable()}}}

\item{columns}{Columns in the data to update on Airtable. Can be a vector of
character strings, unquoted column names, or a tidyselect helper like
\code{\link[tidyselect:starts_with]{tidyselect::starts_with()}}, \code{\link[tidyselect:starts_with]{tidyselect::ends_with()}} or
\code{\link[tidyselect:everything]{tidyselect::everything()}}. Defaults to \code{\link[tidyselect:everything]{tidyselect::everything()}}}

\item{safely}{If \code{TRUE}, confirm number and names of columns to update and
number of rows before executing update.}

\item{return_json}{If \code{TRUE}, return JSON repsponse from the Airtable web API
as a list. If \code{FALSE} (default), return input data.}

\item{parallel}{If \code{TRUE} use parallel processing for encoding large tables.
Currently not implemented in development version.}

\item{...}{
  Arguments passed on to \code{\link[=airtable_request]{airtable_request}}
  \describe{
    \item{\code{url}}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \code{\link[=airtable_request]{airtable_request()}} returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any additional value provided to url is ignored. For
the metadata API (where only a base ID is required) include \code{require_base = TRUE} to pass to \code{\link[=check_airtable_api_url]{check_airtable_api_url()}} when creating a request.}
    \item{\code{base}}{Airtable base id starting with with "app". Optional if url or
airtable are supplied. base and table are both required if url and airtable
are \code{NULL}.}
    \item{\code{table}}{Airtable table id or name. If table is a table ID it is a string
starting with "viw".}
  }}
}
\value{
A data.frame (invisibly) of the input data, to be stored as an object
or piped into further additional functions.
}
\description{
Update one or more columns of data in an Airtable table. Supports batch
updates and parallel JSON encoding (recommended for large tables).
}
