% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_airtable}
\alias{validate_airtable}
\title{Validate a new airtable object}
\usage{
validate_airtable(airtable_obj)
}
\arguments{
\item{airtable_obj}{An object of class `airtable`}
}
\description{
Validate a new airtable object
}
