% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable.R
\name{airtable}
\alias{airtable}
\title{Create a new airtable object}
\usage{
airtable(
  table = NULL,
  base = NULL,
  view = NULL,
  fields = list(),
  api_url = NULL,
  api_version = NULL,
  require_url = TRUE,
  require_fields = FALSE
)
}
\arguments{
\item{table}{Airtable table ID or name. Table ID values start with "tbl". If
table is an Airtable URL matching the pattern
"https://airtable.com/{baseID}/{tableIdOrName}/{viewId}" the URL is parsed
to supply the values for base and table.}

\item{base}{Airtable base ID. Base ID values start with "app". Optional if
table is a URL containing a base ID or required otherwise.}

\item{view}{Optional Airtable view ID. View ID values starts with "viw".}

\item{fields}{An optional \code{airtable_fields_schema} object. Automatically
populated in tables created by \code{airtable_base()}.}

\item{api_url}{Airtable API URL, If \code{NULL} (default), the api_url is set to
getOption("rairtable.api_url", "https://api.airtable.com").}

\item{api_version}{Airtable API version number, If \code{NULL} (default), the
api_version is set to getOption("rairtable.api_version", 0)).}

\item{require_url}{If \code{TRUE} (default), check that x includes a "request_url"
attribute that is a valid Airtable API url.}

\item{require_fields}{If \code{TRUE}, check that x includes a "fields" value that
is a list with the \code{airtable_fields_schema} class. Defaults to \code{FALSE}.}
}
\value{
An \code{airtable} class list of table and fields stored as values of the
list and the base, view, and request_url stored as attributes.
}
\description{
Creates an S3 airtable object, which serves as a pointer for rairtable
functions
}
\examples{
\dontrun{
table <- airtable("Table 1", "appXXXXXXXXXXXXX")
}
}
