% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable_base.R
\name{airtable_base}
\alias{airtable_base}
\alias{get_base_schema}
\alias{get_table_models}
\alias{list_base_tables}
\title{Get metadata for an Airtable base}
\usage{
airtable_base(base = NULL, token = NULL, ...)

get_base_schema(
  base = NULL,
  token = NULL,
  simplifyVector = TRUE,
  ...,
  call = caller_env()
)

get_table_models(
  base = NULL,
  tables = NULL,
  schema = NULL,
  token = NULL,
  ...,
  call = caller_env()
)
}
\arguments{
\item{base}{An Airtable base ID, an airtable object containing a base ID, or
a URL that can be parsed for a base ID. Optional if url or airtable are
supplied to the additional parameters passed to \code{\link[=req_airtable]{req_airtable()}}.}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{...}{Additional parameters:
\itemize{
\item For \code{\link[=airtable_base]{airtable_base()}} and \code{\link[=get_base_schema]{get_base_schema()}}, additional parameters
may include airtable or url. One of the three is required if base is \code{NULL}.
\item For \code{\link[=get_table_models]{get_table_models()}}, additional parameters are passed to
\code{\link[=get_base_schema]{get_base_schema()}} so may include simplifyVector.
}}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{tables}{Optional character vector of table names or IDs to include in
the returned data frame or list.}

\item{schema}{A base schema list from \code{\link[=get_base_schema]{get_base_schema()}} using
\code{simplifyVector = FALSE}. Optional if base is supplied.}
}
\value{
\itemize{
\item \code{\link[=airtable_base]{airtable_base()}} returns an \code{airtable_base_schema} object.
\item \code{\link[=get_base_schema]{get_base_schema()}} returns the response from the Airtable get base schema
API method
\item \code{\link[=get_table_models]{get_table_models()}} returns a tibble with the list of tables
from the schema response or a list.
}
}
\description{
Get an "airtable_base_schema" object, a base schema, or a list of table
models from a base schema.
}
\examples{
\dontrun{
base <- airtable_base("appXXXXXXXXXXXXX")

# read 'table1' from the base
tbl_data <- read_airtable(base$table1)

# view base schema
print(base$schema)
}
}
