% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable_base.R
\name{airtable_base}
\alias{airtable_base}
\title{Title}
\usage{
airtable_base(
  base_id,
  metadata_api_url_pattern = "https://api.airtable.com/v0/meta/bases/\%s/tables"
)
}
\arguments{
\item{base_id}{A valid Airtable base ID}

\item{metadata_api_url_pattern}{URL for GET request. By default 'https://api.airtable.com/v0/meta/bases/%s/tables'}
}
\value{
an `airtable_base_schema` object
}
\description{
Title
}
\examples{
\dontrun{
base <- airtable_base("appXXXXXXXXXXXXX")

# read 'table1' from the base
tbl_data <- read_airtable(base$table1)

# view base schema
print(base$schema)
}
}
