% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable_base.R
\name{airtable_base}
\alias{airtable_base}
\alias{list_airtable_bases}
\title{Get metadata for an Airtable base or list bases for an account}
\usage{
airtable_base(
  base = NULL,
  ...,
  metadata_api_url_pattern = deprecated(),
  as_tibble = FALSE,
  token = NULL
)

list_airtable_bases(..., base = NULL, token = NULL)
}
\arguments{
\item{base}{A character vector of base id values. If supplied to
\code{\link[=list_airtable_bases]{list_airtable_bases()}}, return only those Airtable bases with matching
IDs.}

\item{...}{Additional parameters passed to \code{\link[=req_airtable_schema]{req_airtable_schema()}}. Supports
the use of an optional url or airtable parameter instead of base for
\code{\link[=airtable_base]{airtable_base()}}. Additional parameters are not used by
\code{\link[=list_airtable_bases]{list_airtable_bases()}} at present.}

\item{metadata_api_url_pattern}{Metadata API URL pattern. Deprecated.}

\item{as_tibble}{If \code{TRUE}, return a tibble with a row for each table list
column of fields. Defaults to \code{FALSE}.}

\item{token}{Airtable personal access token or API key.}
}
\value{
An \code{airtable_base_schema} object or a tibble listing the tables from
the Airtable base.
}
\description{
Get an "airtable_base_schema" object, a tibble with a list of tables, or a
tibble with a list of bases associated with the provided personal access
token.
}
\examples{
\dontrun{
base <- airtable_base("appXXXXXXXXXXXXX")

# read 'table1' from the base
tbl_data <- read_airtable(base$table1)

# view base schema
print(base$schema)
}
}
