% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable_base.R
\name{airtable_base}
\alias{airtable_base}
\alias{list_bases}
\title{Get metadata for an Airtable base or list bases for an account}
\usage{
airtable_base(
  base = NULL,
  ...,
  metadata_api_url_pattern = deprecated(),
  as_tibble = FALSE,
  token = NULL
)

list_bases(base = NULL, token = NULL)
}
\arguments{
\item{base}{An Airtable base ID. For \code{\link[=airtable_base]{airtable_base()}}, base is optional if
url or airtable are supplied and must be a string. For \code{\link[=list_bases]{list_bases()}}, base
is a character vector of base id values and the function returns only those
Airtable bases with matching IDs.}

\item{...}{Optional airtable or url parameter. url must be a URL for an
Airtable API call or an Airtable base. url is optional if airtable or base
are supplied and airtable is optional if base or url are supplied.}

\item{metadata_api_url_pattern}{Metadata API URL pattern. Deprecated.}

\item{as_tibble}{If \code{TRUE}, return a tibble with a row for each table list
column of fields. Defaults to \code{FALSE}.}

\item{token}{Airtable personal access token or API key.}
}
\value{
An \code{airtable_base_schema} object or a tibble listing the tables from
the Airtable base.
}
\description{
Get an "airtable_base_schema" object, a tibble with a list of tables, or a
tibble with a list of bases associated with the provided personal access
token. Additional parameters are not used by \code{\link[=list_bases]{list_bases()}} at present.
}
\examples{
\dontrun{
base <- airtable_base("appXXXXXXXXXXXXX")

# read 'table1' from the base
tbl_data <- read_airtable(base$table1)

# view base schema
print(base$schema)
}
}
