% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable_fields_schema.R
\name{airtable_fields_schema}
\alias{airtable_fields_schema}
\title{Create an \code{airtable_fields_schema} objec}
\usage{
airtable_fields_schema(field_schema)
}
\arguments{
\item{field_schema}{A list of lists. Each child list should have a \code{name},
\code{type}, and \code{id} as well as an optional \code{options}}
}
\value{
An \code{airtable_fields_schema} object
}
\description{
Create an \code{airtable_fields_schema} object from a list of lists to use for data
validation.
}
