% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable_url.R
\name{airtable_url}
\alias{airtable_url}
\alias{is_airtable_url}
\alias{parse_airtable_url}
\title{Test and parse Airtable URLs}
\usage{
is_airtable_url(url, base_url = NULL, allow_shared = FALSE)

parse_airtable_url(
  url,
  base_url = NULL,
  api_url = NULL,
  api_version = NULL,
  table_name = NULL,
  view_name = NULL,
  require_table = FALSE,
  require_view = FALSE,
  require_field = FALSE,
  call = caller_env()
)
}
\arguments{
\item{url}{Airtable URL or API URL, typically in the form
"https://airtable.com/{baseID}/{tableIdOrName}/{viewId}" or
"https://api.airtable.com/v0/{baseId}/{tableIdOrName}". Note that, by
default, shared URLs of the form "https://airtable.com/shr..." are not
supported.}

\item{base_url}{Base URL for an Airtable base or view. Defaults \code{NULL} and
set to \code{getOption("rairtable.base_url", "https://airtable.com")}.}

\item{allow_shared}{If \code{FALSE} (default), \code{\link[=is_airtable_url]{is_airtable_url()}} returns \code{FALSE}
for any shared base or view URL that starts with
"https://airtable.com/shr".}

\item{api_url}{Airtable API URL, If \code{NULL} (default), the api_url is set to
\code{getOption("rairtable.api_url", "https://api.airtable.com")}.}

\item{api_version}{Airtable API version number, If \code{NULL} (default), the
api_version is set to \verb{getOption("rairtable.api_version", 0))}.}

\item{table_name, view_name}{Pattern to use when parsing a table name or view
name. Defaults to \code{NULL} which extracts strings starting with "tbl" for
tables IDs and "viw" for view IDs.}

\item{require_table}{If \code{TRUE}, the string "tbl" must be included in the
supplied url. If require_table is a string, it is treated as the required
table name which must be part of url. If \code{FALSE} (default), there is no
check for a table being present in the API request URL.}

\item{require_view}{If \code{TRUE}, the string "viw" must be included in the
string supplied to url. If \code{FALSE} (default), there is no check
for a view being present in the API request URL.}

\item{require_field}{If \code{TRUE}, the url supplied to \code{\link[=parse_airtable_url]{parse_airtable_url()}}
must include a string starting with "fld" to use as the field ID .}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Utility functions primarily intended for use by developers for checking and
validating Airtable URLs, \code{\link[=is_airtable_url]{is_airtable_url()}} returns \code{TRUE} if a url starts
with the base_url value. \code{\link[=parse_airtable_url]{parse_airtable_url()}} accepts a base or API url as
an input and returns a named list with the base, table, view, and field ID
values found in the URL.
}
\details{
For more information on the expected URL format for different Airtable API
end points, see the web API documentation at
\url{https://airtable.com/developers/web/api/introduction}
}
