% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_field.R
\name{create_field}
\alias{create_field}
\alias{update_field}
\title{Create a field or column in an Airtable table}
\usage{
create_field(
  airtable = NULL,
  name = NULL,
  type = "singleLineText",
  description = NULL,
  options = NULL,
  field = NULL,
  token = NULL,
  ...
)

update_field(
  airtable = NULL,
  column = NULL,
  name = NULL,
  description = NULL,
  token = NULL,
  ...
)
}
\arguments{
\item{airtable}{An airtable class object created with \code{\link[=airtable]{airtable()}}. airtable
must include a request_url attribute. url is ignored if airtable is
provided. Optional if url or base \emph{and} table are supplied.}

\item{name}{Name for the field or column to create.}

\item{type}{Field type, defaults to 'singleLineText'. See
\code{field_types} for list of supported field types.}

\item{description}{Optional. Description for the field or column to create or
description to add to updated field. Description must be no longer than
20,000 characters.}

\item{options}{Named list of field options. Required for some field types.
See the Airtable Web API documentation on field types and cell values for
more information: \url{https://airtable.com/developers/web/api/field-model}
Default: \code{NULL}.}

\item{field}{A named list that can be used to set the name, type,
description, or options of a field to create. If field is supplied with a
type, any type value is ignored. Any supplied name, description, or options
are used if provided.}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{...}{For \code{\link[=request_airtable]{request_airtable()}}, additional parameters passed to
\code{\link[=parse_airtable_url]{parse_airtable_url()}} if an Airtable url is provided. For
\code{\link[=req_airtable]{req_airtable()}}, additional parameters passed to
\code{\link[httr2:req_template]{httr2::req_template()}} if template is not \code{NULL} or
\code{\link[httr2:req_url]{httr2::req_url_query()}} if template is \code{NULL}.}

\item{column}{Field ID of column to update.}
}
\value{
A list with the created field name, field ID, description (if
provided), and options.
}
\description{
Create a field or column in the table of an Airtable base. Find more
information on the create field API method and field types and options in the
Airtable Web API documentation:
\url{https://airtable.com/developers/web/api/create-field} Both name and type are
required unless field is provided.
}
