% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_field.R
\name{create_field}
\alias{create_field}
\title{Create a field or column in an Airtable table}
\usage{
create_field(
  airtable = NULL,
  name,
  type = "singleLineText",
  description = NULL,
  options = NULL,
  token = NULL,
  ...
)
}
\arguments{
\item{airtable}{An airtable class object created with \code{\link[=airtable]{airtable()}}. airtable
must include a request_url attribute. url is ignored if airtable is
provided.  Optional if url or base \emph{and} table are supplied.}

\item{name}{Name for the field or column to create. Required.}

\item{type}{Field type, defaults to 'singleLineText'. Required. See
field_types for list of supported field types.}

\item{description}{Optional. Description for the field or column to create.
Default: \code{NULL}}

\item{options}{Named list of field options. Required for some field types.
See the Airtable Web API documentation on field types and cell values for
more information: \url{https://airtable.com/developers/web/api/field-model}
Default: \code{NULL}}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{...}{
  Arguments passed on to \code{\link[=req_airtable_schema]{req_airtable_schema}}
  \describe{
    \item{\code{url}}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \code{\link[=airtable_request]{airtable_request()}} returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any additional value provided to url is ignored. For
the metadata API (where only a base ID is required) include \code{require_base = TRUE} to pass to \code{\link[=check_airtable_api_url]{check_airtable_api_url()}} when creating a request.}
    \item{\code{base}}{Airtable base id starting with with "app". Optional if url or
airtable are supplied or if require_base is \code{FALSE}. base and table are
both required if url and airtable are \code{NULL} and require_base is \code{TRUE}.}
    \item{\code{table}}{Airtable table id or name. Table ID values are strings starting
with "tbl". Optional if require_table is \code{FALSE}.}
  }}
}
\value{
A list with the created field name, field ID, description (if
provided), and options.
}
\description{
Create a field or column in the table of an Airtable base. Find more
information on the create field API method and field types and options in the
Airtable Web API documentation:
\url{https://airtable.com/developers/web/api/create-field}
}
