% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\alias{update_table}
\alias{make_table_model}
\title{Create or update an Airtable table}
\usage{
create_table(
  airtable = NULL,
  model = NULL,
  name = NULL,
  description = NULL,
  fields = NULL,
  token = NULL,
  ...
)

update_table(airtable = NULL, name = NULL, description = NULL, ...)

make_table_model(
  name = NULL,
  description = NULL,
  fields = NULL,
  ...,
  call = caller_env()
)
}
\arguments{
\item{airtable}{An airtable class object created with \code{\link[=airtable]{airtable()}}. airtable
must include a request_url attribute. url is ignored if airtable is
provided. Optional if url or base \emph{and} table are supplied.}

\item{model}{A table model with a name, description, and field list.
Typically created with \code{\link[=make_table_model]{make_table_model()}}.}

\item{name}{Table name. Required to create model for \code{\link[=create_table]{create_table()}}.}

\item{description}{Table description. Optional.}

\item{fields}{Field list. Required to create model for \code{\link[=create_table]{create_table()}}.}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{...}{Additional parameters passed to \code{\link[=req_airtable]{req_airtable()}}}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
Invisibly returns a list with the table ID, name, description,
primary field ID, and field names, types, and IDs.
}
\description{
Create or update an Airtable table or create a table model. \code{\link[=create_table]{create_table()}}
requires both a name and fields (or a model with a name and fields). Either a
name or description must be provided for \code{\link[=update_table]{update_table()}}.
}
\details{
Find more information on the create table API:
\url{https://airtable.com/developers/web/api/create-table} or the update table
API: \url{https://airtable.com/developers/web/api/update-table} Find more
information about table models for the Airtable API:
\url{https://airtable.com/developers/web/api/model/table-model}
}
\examples{
\dontrun{
if (interactive()) {
  create_table(
    base = "<exmaple base ID>",
    name = "Apartments",
    description = "A to-do list of places to visit",
    fields = list(
      list(
        "description" = "Name of the apartment",
        "name" = "Name",
        "type" = "singleLineText"
      ),
      list(
        "name" = "Address",
        "type" = "singleLineText"
      ),
      list(
        "name" = "Visited",
        "options" = list(
          "color" = "greenBright",
          "icon" = "check"
        ),
        "type" = "checkbox"
      )
    )
  )
}
}
}
