% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{field_types}
\alias{field_types}
\title{Airtable Field Types}
\format{
A data frame with 33 rows and 6 variables:
\describe{
\item{\code{type}}{Field type string}
\item{\code{name}}{Name of field type}
\item{\code{url}}{API reference URL}
\item{\code{class}}{Correponding class}
\item{\code{read_only}}{Flag for read-only field types}
\item{\code{options}}{Flag for field types with options}
}
}
\usage{
field_types
}
\description{
A data frame of Airtable field types as of 2023-06-27 Find more information
on Airtable field types at:
\url{https://airtable.com/developers/web/api/field-model}
}
\keyword{datasets}
