% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_records.R
\name{insert_records}
\alias{insert_records}
\title{Insert records into an Airtable table}
\usage{
insert_records(data, airtable, parallel = FALSE, batch_size = 10)
}
\arguments{
\item{data}{A dataframe containing records to insert}

\item{airtable}{An airtable object}

\item{parallel}{Use parallel processing for encoding large tables}

\item{batch_size}{Number of records per request to insert. Maximum of 10}
}
\description{
Insert records into an Airtable table
}
