% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_airtable_obj.R
\name{is_airtable_obj}
\alias{is_airtable_obj}
\alias{check_airtable_obj}
\title{Test or check if an object is an airtable object}
\usage{
is_airtable_obj(x)

check_airtable_obj(
  x,
  require_url = TRUE,
  require_table = TRUE,
  require_view = FALSE,
  require_fields = FALSE,
  allow_null = FALSE,
  call = caller_env()
)
}
\arguments{
\item{x}{Object to test or check.}

\item{require_url}{If \code{TRUE} (default), check that x includes a "request_url"
attribute that is a valid Airtable API url.}

\item{require_table}{If \code{TRUE}, the string "tbl" must be included in the
supplied url. If a string with a table name is provided, the string must be
part of the string supplied to url. If \code{FALSE} (default), there is no check
for a table being present in the URL.}

\item{require_view}{If \code{TRUE}, the string "viw" must be included in the
string supplied to url. If \code{FALSE} (default), there is no check
for a view being present in the URL.}

\item{require_fields}{If \code{TRUE}, check that x includes a "fields" value that
is a list with the \code{airtable_fields_schema} class. Defaults to \code{FALSE}.}

\item{allow_null}{If \code{FALSE} (default) and url is \code{NULL},
\code{\link[=check_airtable_url]{check_airtable_url()}} errors. If \code{TRUE}, a \code{NULL} value for url is
allowed.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=is_airtable_obj]{is_airtable_obj()}} returns TRUE for airtable class objects or FALSE
otherwise. \code{\link[=check_airtable_obj]{check_airtable_obj()}} errors if the input is not an airtable
object with the option to ignore \code{NULL} inputs. This check can also
optionally require an airtable object to include a request URL, a table, and
a view. Note that this is an internal function primarily intended for
developer use.
}
\keyword{internal}
