% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bases.R
\name{list_bases}
\alias{list_bases}
\title{List Airtable bases accessible with a personal access token}
\usage{
list_bases(bases = NULL, token = NULL, simplifyVector = TRUE, ...)
}
\arguments{
\item{bases}{A character vectors of base ID or name values. If provided, the
results of list_bases are filtered to matching bases only. Defaults to
\code{NULL} which returns all bases associated with the personal access token or
API key.}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{...}{Other arguments passed on to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} and
\code{\link[xml2:read_xml]{xml2::read_xml()}} respectively.}
}
\value{
A tibble or list of base IDs, names, and permission levels.
}
\description{
Use the Airtable Metadata API to list all Airtable bases accessible with a
personal access token. Optionally filter bases by name or base ID.
}
\seealso{
\code{\link[=get_table_models]{get_table_models()}}
}
