% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_config.R
\name{make_field_config}
\alias{make_field_config}
\title{Make a field config or a list of field config elements}
\usage{
make_field_config(
  field = NULL,
  name = NULL,
  type = NULL,
  ...,
  call = caller_env()
)
}
\arguments{
\item{field}{A list with a name and type value. Optionally can include
options and description. If field is a single row data frame, it is
converted to a list. If field is a list of lists, the function is applied
to each element in the list.}

\item{name}{A string to use as a name for the field. Optional if field is supplied.}

\item{type}{A type to use for the field. Must be a string matching a value in
\code{field_types}. Optional if field is supplied.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=make_field_config]{make_field_config()}} is a helper function for creating a field config for
adding new fields to an existing table or to add to a table config object.
}
\details{
Learn more about the Airtable Web API field model:
\url{https://airtable.com/developers/web/api/field-model}
}
\seealso{
field_types
}
