% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{post}
\alias{post}
\title{Send batch post request}
\usage{
post(records, airtable_obj)
}
\arguments{
\item{records}{JSON records to post}

\item{airtable_obj}{An object of class `airtable`}
}
\description{
Send batch post request
}
