% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rairtable-package.R
\docType{package}
\name{rairtable-package}
\alias{rairtable}
\alias{rairtable-package}
\title{rairtable: Efficient Wrapper for the 'Airtable' API}
\description{
Efficient CRUD interface for the 'Airtable' API \url{https://airtable.com/developers/web/api}, supporting batch requests and parallel encoding of large data sets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matthewjrogers.github.io/rairtable/}
  \item Report bugs at \url{https://github.com/matthewjrogers/rairtable/issues}
}

}
\author{
\strong{Maintainer}: Matthew Rogers \email{matthew.rogers09@gmail.com}

}
\keyword{internal}
