% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_airtable.R
\name{read_airtable}
\alias{read_airtable}
\alias{list_records}
\alias{get_record}
\title{Read records from an Airtable table}
\usage{
read_airtable(
  airtable = NULL,
  fields = NULL,
  id_to_col = TRUE,
  airtable_id_col = NULL,
  max_rows = deprecated(),
  token = NULL,
  ...
)

list_records(
  airtable = NULL,
  view = NULL,
  sort = NULL,
  max_records = 100,
  page_size = NULL,
  tz = NULL,
  locale = NULL,
  fields_by_id = FALSE,
  fields = NULL,
  desc = FALSE,
  cell_format = NULL,
  token = NULL,
  offset = NULL,
  ...
)

get_record(airtable = NULL, record, airtable_id_col = NULL, token = NULL, ...)
}
\arguments{
\item{airtable}{An \code{airtable} class object. Optional for \code{\link[=read_airtable]{read_airtable()}} if
url is supplied. For \code{\link[=list_records]{list_records()}} and
\code{\link[=get_record]{get_record()}}, support the airtable, url, or a base \emph{and} table
parameter.}

\item{fields}{Fields to return from Airtable base, Default: \code{NULL}}

\item{id_to_col}{If \code{TRUE} (default), the airtable record IDs will be added
to the returned data frame as a new column. If \code{FALSE}, the airtable record
IDs are used as row names.}

\item{airtable_id_col}{Airtable record ID column name assigned to returned
data.frame. Ignored if id_to_col is \code{TRUE}. Defaults to \code{NULL} which is set
to \code{getOption("rairtable.id_col", "airtable_record_id")}.}

\item{max_rows}{Optional maximum number of rows to read. Defaults to \code{NULL}}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{...}{For \code{\link[=read_airtable]{read_airtable()}}, additional query parameters can be passed
to \code{\link[=req_query_airtable]{req_query_airtable()}}. For \code{\link[=list_records]{list_records()}} and
\code{\link[=get_record]{get_record()}}, additional parameters, such as url, can be passed
to \code{\link[=airtable_request]{airtable_request()}}.}

\item{view}{Airtable view ID or name, Default: \code{NULL}. If the supplied url or
airtable object includes a view, the view provided to this parameter is
ignored.}

\item{sort}{Field to sort by, Default: \code{NULL}}

\item{max_records}{Maximum number of records to return, Default: \code{NULL}. Must
be 100 or less.}

\item{page_size}{Max records to return per page, Default: \code{NULL}}

\item{tz, locale}{Time zone and locale, Defaults to \code{NULL}. If cell_format is
"string", tz defaults to \code{Sys.timezone()} and locale defaults to
\code{Sys.getlocale("LC_TIME")}.}

\item{fields_by_id}{If \code{TRUE}, return fields by id, Default: \code{FALSE}}

\item{desc}{If \code{TRUE}, sort in descending order, Default: \code{FALSE}}

\item{cell_format}{Cell format for "Link to another record" fields. Defaults
to "json" which returns a unique record ID. A "string" cell_format returns
the displayed character string.}

\item{offset}{Offset value. Primarily intended for internal use.}

\item{record}{Record ID number. Required for \code{\link[=get_record]{get_record()}}.}
}
\value{
A data.frame with the records from the Airtable base and table
provided by the airtable parameter.
}
\description{
Read records from a table in an Airtable base. \code{\link[=read_airtable]{read_airtable()}} supports
basic access to data based on an airtable object. \code{\link[=list_records]{list_records()}}
and \code{\link[=get_record]{get_record()}} allows you to pass url or base and table directly
and supports more options from the Airtable API.
}
\details{
Find more information on the list records API method:
\url{https://airtable.com/developers/web/api/list-records} or get record API
method: \url{https://airtable.com/developers/web/api/get-record}
}
