% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_airtable.R
\name{read_airtable}
\alias{read_airtable}
\title{Read table from Airtable}
\usage{
read_airtable(airtable, id_to_col = FALSE, max_rows = 50000)
}
\arguments{
\item{airtable}{An airtable object}

\item{id_to_col}{If TRUE, store airtable ID as a column rather than as row names}

\item{max_rows}{Optional maximum number of rows to read}
}
\description{
Connect to and read values from an Airtable table.
}
