% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_records.R
\name{req_delete_records}
\alias{req_delete_records}
\alias{req_delete_record}
\title{Delete one or more records from an Airtable table}
\usage{
req_delete_records(req = NULL, ..., records, token = NULL, call = caller_env())

req_delete_record(req = NULL, ..., record, token = NULL, call = caller_env())
}
\arguments{
\item{req}{If req is supplied, url or any additional parameters passed to ...
are ignored.}

\item{...}{
  Arguments passed on to \code{\link[=request_airtable]{request_airtable}}
  \describe{
    \item{\code{airtable}}{An airtable class object created with \code{\link[=airtable]{airtable()}}. airtable
must include a request_url attribute. url is ignored if airtable is
provided. Optional if url or base \emph{and} table are supplied.}
    \item{\code{url}}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \code{\link[=request_airtable]{request_airtable()}} returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any supplied url is ignored.}
    \item{\code{base}}{Airtable base id starting with with "app". Optional if url or
airtable are supplied or if require_base is \code{FALSE}. base and table are
both required if url and airtable are \code{NULL} and require_base is \code{TRUE}.}
    \item{\code{table}}{Airtable table id or name. Table ID values are strings starting
with "tbl". Optional if require_table is \code{FALSE}.}
    \item{\code{view}}{Airtable view ID. View ID values starts with "viw". Optional if
require_view is \code{FALSE}.}
    \item{\code{api_url}}{Airtable API URL, If \code{NULL} (default), the api_url is set to
\code{getOption("rairtable.api_url", "https://api.airtable.com")}.}
    \item{\code{api_version}}{Airtable API version number, If \code{NULL} (default), the
api_version is set to \verb{getOption("rairtable.api_version", 0))}.}
    \item{\code{require_base}}{If \code{TRUE} (default), the string "app" must be included in
the supplied API request URL.}
    \item{\code{require_table}}{If \code{TRUE}, the string "tbl" must be included in the
supplied url. If require_table is a string, it is treated as the required
table name which must be part of url. If \code{FALSE} (default), there is no
check for a table being present in the API request URL.}
    \item{\code{require_view}}{If \code{TRUE}, the string "viw" must be included in the
string supplied to url. If \code{FALSE} (default), there is no check
for a view being present in the API request URL.}
  }}

\item{records}{Required. Character vector with record IDs. If only one record
is provided, the parameters are passed to \code{\link[=req_delete_record]{req_delete_record()}}
https://airtable.com/developers/web/api/delete-multiple-records.}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{batch}{If \code{TRUE} (default) and more than 10 records are provided, the
records vector is split into batches of 10. If batch is \code{FALSE}, records
must be length 10 or less. This default batch size is set by
"rairtable.batch_size" option. The option default is 10 may be changed if
the API limit is modified in the future.}
}
\description{
Delete one record from an Airtable table with \code{\link[=req_delete_record]{req_delete_record()}} or delete
multiple records with \code{\link[=req_delete_records]{req_delete_records()}}.
}
\keyword{internal}
