% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_records.R
\name{req_delete_records}
\alias{req_delete_records}
\alias{req_delete_record}
\title{Delete one or more records from an Airtable table}
\usage{
req_delete_records(req = NULL, ..., records, token = NULL, call = caller_env())

req_delete_record(req = NULL, ..., record, token = NULL, call = caller_env())
}
\arguments{
\item{req}{If req is supplied, url or any additional parameters passed to ...
are ignored.}

\item{...}{
  Arguments passed on to \code{\link[=airtable_request]{airtable_request}}
  \describe{
    \item{\code{url}}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \code{\link[=airtable_request]{airtable_request()}} returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any additional value provided to url is ignored. For
the metadata API (where only a base ID is required) include \code{require_base = TRUE} to pass to \code{\link[=check_airtable_api_url]{check_airtable_api_url()}} when creating a request.}
    \item{\code{base}}{Airtable base id starting with with "app". Optional if url or
airtable are supplied. base and table are both required if url and airtable
are \code{NULL}.}
    \item{\code{table}}{Airtable table id or name. If table is a table ID it is a string
starting with "viw".}
    \item{\code{airtable}}{An airtable class object created with \code{\link[=airtable]{airtable()}}. airtable
must include a request_url attribute. url is ignored if airtable is
provided.  Optional if url or base \emph{and} table are supplied.}
    \item{\code{api_url}}{Airtable API URL, If \code{NULL} (default), the api_url is set to
\code{getOption("rairtable.api_url", "https://api.airtable.com")}.}
    \item{\code{api_version}}{Airtable API version number, If \code{NULL} (default), the
api_version is set to \verb{getOption("rairtable.api_version", 0))}.}
  }}

\item{records}{Required. Character vector with record IDs. If only one record
is provided, the parameters are passed to \code{\link[=req_delete_record]{req_delete_record()}}
https://airtable.com/developers/web/api/delete-multiple-records.}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{batch}{If \code{TRUE} (default) and more than 10 records are provided, the
records vector is split into batches of 10. If batch is \code{FALSE}, records
must be length 10 or less. This default batch size is set by
"rairtable.batch_size" option. The option default is 10 may be changed if
the API limit is modified in the future.}
}
\description{
Delete one record from an Airtable table with \code{\link[=req_delete_record]{req_delete_record()}} or delete
multiple records with \code{\link[=req_delete_records]{req_delete_records()}}.
}
\details{
More information on the Airtable API:
\itemize{
\item \url{https://airtable.com/developers/web/api/delete-record}
\item \url{https://airtable.com/developers/web/api/delete-multiple-records}
}
}
\keyword{internal}
