% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_airtable.R
\name{req_perform_offset}
\alias{req_perform_offset}
\alias{resp_body_records}
\title{Get the body from API requests with a series of offset values}
\usage{
req_perform_offset(
  req,
  offset = NULL,
  type = "combine",
  airtable_id_col = NULL,
  metadata = c("id", "createdTime", "commentCount"),
  max_rows = NULL,
  .name_repair = "unique",
  call = caller_env()
)

resp_body_records(
  resp,
  simplifyVector = TRUE,
  type = "combine",
  airtable_id_col = NULL,
  metadata = c("id", "createdTime", "commentCount"),
  .name_repair = "unique",
  ...,
  call = caller_env()
)
}
\arguments{
\item{req}{A modified HTTP request from \code{\link[=request_airtable]{request_airtable()}} or a httr2
function.}

\item{offset}{Offset value to passed to \code{\link[httr2:req_url]{httr2::req_url_query()}} as part of
the API call. See
\url{https://airtable.com/developers/web/api/list-records#response-offset} for
more information on how the offset value is used by the Airtable API.
Primarily intended for developer use only.}

\item{type}{"combine" (default) combines the data frames with fields and
records from the API response. Additional supported options are "records"
and "fields".}

\item{airtable_id_col}{An alternate name to use for the id column of the
response data frame.}

\item{metadata}{Columns to return when type is "records" or to combine
with fields when type is "combine".}

\item{.name_repair}{One of "unique" (default), "universal", "check_unique",
"unique_quiet", or "universal_quiet" passed to \code{\link[vctrs:vec_bind]{vctrs::vec_cbind()}}. See
\link{vctrs:vec_as_names} for the meaning of these options.}

\item{...}{Additional parameters passed to \code{\link[vctrs:vec_bind]{vctrs::vec_cbind()}}. Not used.}
}
\description{
Get the body from API requests with a series of offset values
}
\keyword{internal}
