% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_records.R
\name{req_update_records}
\alias{req_update_records}
\alias{req_update_record}
\title{Update one or more records in an Airtable base}
\usage{
req_update_records(
  req = NULL,
  ...,
  records,
  data,
  fields_by_id = FALSE,
  typecast = FALSE,
  method = NULL,
  token = NULL,
  call = caller_env()
)

req_update_record(
  req = NULL,
  ...,
  record,
  data,
  typecast = TRUE,
  method = NULL,
  token = NULL,
  call = caller_env()
)
}
\arguments{
\item{req}{A HTTP response created by \code{\link[=req_airtable]{req_airtable()}}. Optional if
airtable, url, \emph{or} base and table are passed to \code{\link[=request_airtable]{request_airtable()}}.}

\item{...}{
  Arguments passed on to \code{\link[=request_airtable]{request_airtable}}
  \describe{
    \item{\code{airtable}}{An airtable class object created with \code{\link[=airtable]{airtable()}}. airtable
must include a request_url attribute. url is ignored if airtable is
provided. Optional if url or base \emph{and} table are supplied.}
    \item{\code{url}}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \code{\link[=request_airtable]{request_airtable()}} returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any supplied url is ignored.}
    \item{\code{base}}{Airtable base id starting with with "app". Optional if url or
airtable are supplied or if require_base is \code{FALSE}. base and table are
both required if url and airtable are \code{NULL} and require_base is \code{TRUE}.}
    \item{\code{table}}{Airtable table id or name. Table ID values are strings starting
with "tbl". Optional if require_table is \code{FALSE}.}
    \item{\code{view}}{Airtable view ID. View ID values starts with "viw". Optional if
require_view is \code{FALSE}.}
    \item{\code{require_base}}{If \code{TRUE} (default), the string "app" must be included in
the supplied API request URL.}
    \item{\code{require_table}}{If \code{TRUE}, the string "tbl" must be included in the
supplied url. If require_table is a string, it is treated as the required
table name which must be part of url. If \code{FALSE} (default), there is no
check for a table being present in the API request URL.}
    \item{\code{require_view}}{If \code{TRUE}, the string "viw" must be included in the
string supplied to url. If \code{FALSE} (default), there is no check
for a view being present in the API request URL.}
  }}

\item{records, record}{Record ID or IDs to update as a character vector.}

\item{data}{Data to include in body.}

\item{method}{Custom HTTP method}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{field_by_id}{If \code{TRUE}, the response to the API request is keyed by
field IDs instead of field names. The response is only returned if
\code{return_data = FALSE}.}
}
\description{
Update one or more records in an Airtable base
}
\keyword{internal}
