% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_airtable_meta.R
\name{request_airtable_meta}
\alias{request_airtable_meta}
\title{Create and perform a request for the Airtable Metadata API}
\usage{
request_airtable_meta(
  url = NULL,
  base = NULL,
  table = NULL,
  column = NULL,
  airtable = NULL,
  ...,
  meta = NULL,
  template = NULL,
  data = NULL,
  token = NULL,
  call = caller_env()
)
}
\arguments{
\item{url}{A URL for an API call or an Airtable view. If url is provided and
is a valid URL, \code{\link[=request_airtable]{request_airtable()}} returns  \code{httr2::request(url)}. url
is optional if airtable \emph{or} base and table are supplied. If url is an
Airtable view, any additional base or table values provided are ignored. If
airtable is provided, any supplied url is ignored.}

\item{base}{Airtable base id starting with with "app". Optional if url or
airtable are supplied or if require_base is \code{FALSE}. base and table are
both required if url and airtable are \code{NULL} and require_base is \code{TRUE}.}

\item{table}{Airtable table id or name. Table ID values are strings starting
with "tbl". Optional if require_table is \code{FALSE}.}

\item{column}{Column ID, Defaults to \code{NULL}. Only required when \code{meta = "update_field"}. Column can be a field URL.}

\item{airtable}{An airtable class object created with \code{\link[=airtable]{airtable()}}. airtable
must include a request_url attribute. url is ignored if airtable is
provided. Optional if url or base \emph{and} table are supplied.}

\item{...}{Additional parameters passed to \code{\link[=req_airtable]{req_airtable()}}}

\item{meta}{Type of metadata API call to use for bases, tables, or fields.
Options include "schema_base", "create_base", "list_base", "update_table",
"create_table", "update_field", or "create_field". Optional if template is
provided.}

\item{template}{URL template passed to \code{\link[=req_airtable]{req_airtable()}} if meta is \code{NULL}.}

\item{data}{Data to include in body.}

\item{token}{Airtable personal access token, Defaults to \code{NULL} which is set
to environmental variable for "AIRTABLE_PAT" (from \code{\link[=get_airtable_pat]{get_airtable_pat()}}) or
"AIRTABLE_API_KEY" (from \code{\link[=get_airtable_api_key]{get_airtable_api_key()}}).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
An HTTP response as an S3 list with class httr2_request.
}
\description{
Create a request to get metadata or update metadata for bases, tables, or
fields using the Airtable metadata API. Note that the metadata API does \emph{not}
support the legacy Airtable API key so a personal access token is required.
Save a personal access token with \code{\link[=set_airtable_pat]{set_airtable_pat()}}. This function does
not currently support the API for API for views or permissions which are both
limited to Enterprise user accounts.
}
\keyword{internal}
