% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_airtable_api_key.R
\name{set_airtable_api_key}
\alias{set_airtable_api_key}
\title{Set or install Airtable API key}
\usage{
set_airtable_api_key(key, install = FALSE)
}
\arguments{
\item{key}{A valid Airtable API key}

\item{install}{Add your API key to .Renviron for future sessions. Optionally overwrite an existing Airtable API key.}
}
\value{
No return value, called for side effects
}
\description{
Set Airtable API key as an environment variable, and optionally install the API key to your .Renviron file for future use.
}
\examples{
\dontrun{
airtable_api_key("XXXXXXXXXX", install = TRUE)
}

}
