% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_records.R
\name{update_records}
\alias{update_records}
\title{Update airtable records}
\usage{
update_records(
  data,
  airtable,
  columns = dplyr::everything(),
  airtable_id_col = NULL,
  safely = TRUE,
  parallel = TRUE,
  batch_size = 10
)
}
\arguments{
\item{data}{A dataframe containing the records and fields to update}

\item{airtable}{An airtable object}

\item{columns}{Columns in the data to update on Airtable. Defaults to `dplyr::everything()`}

\item{airtable_id_col}{Column containing Airtable record IDs. Not required if record IDs are stored in row names as returned from `read_airtable`}

\item{safely}{If `TRUE`, confirm number and names of columns to update and number of rows befor executing update.}

\item{parallel}{If `TRUE` use parallel processing for encoding large tables}

\item{batch_size}{Number of records to update per request. Maximum of 10}
}
\description{
Update airtable records
}
